// SMSfrontEnd/src/app/models/customer.ts
// export interface Customer {
  export class Customer {
    id: number;
    customerName: string;
    idNumber: string;
    tinNo: string;
    contactPerson: string;
    custContact: string;
    country: string;
    region: string;
    city: string;
    remarks: string;
    vatNo: string;
    customerType: string;
    businessType: string;
    customerStatus: string;
    branchId: number;
    phoneNo: string; 
    paymentType: string;
  
  }